chore: silent a warning on integer type conversion

+++ macos_keycode.h
@@ -233,7 +233,7 @@
 
 // conversion functions
 
-int osx_modifiers_to_rime_modifiers(unsigned modifiers);
+int osx_modifiers_to_rime_modifiers(unsigned long modifiers);
 int osx_keycode_to_rime_keycode(int keycode, int keychar, int shift, int caps);
